requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-pgcrypto-fields==2.6.0
django-otp==1.3.0
django-axes==6.1.1
drf-spectacular==0.26.5
django-filter==23.3
django-cors-headers==4.3.0
Faker==20.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
boto3==1.34.0
django-storages==1.14.2
whitenoise==6.6.0
